Split exporters
