+++ locations/exporters/geojson.py
@@ -5,7 +5,7 @@ import json
 import logging
 import uuid
 
-from scrapy.exporters import JsonItemExporter, JsonLinesItemExporter
+from scrapy.exporters import JsonItemExporter
 from scrapy.utils.misc import walk_modules
 from scrapy.utils.python import to_bytes
 from scrapy.utils.spider import iter_spider_classes
@@ -90,39 +90,6 @@ def get_dataset_attributes(spider_name) -> {}:
     dataset_attributes["@spider"] = spider_name
 
     return dataset_attributes
-
-
-class LineDelimitedGeoJsonExporter(JsonLinesItemExporter):
-    dataset_attributes = None
-    first_item = True
-
-    def export_item(self, item):
-        if self.first_item:
-            self.first_item = False
-            self.dataset_attributes = get_dataset_attributes(item["extras"].get("@spider"))
-        super().export_item(item)
-
-    def _get_serialized_fields(self, item, default_value=None, include_empty=None):
-        feature = []
-        feature.append(("type", "Feature"))
-        feature.append(("id", compute_hash(item)))
-        feature.append(("dataset_attributes", self.dataset_attributes))
-        feature.append(("properties", item_to_properties(item)))
-
-        lat = item.get("lat")
-        lon = item.get("lon")
-        geometry = item.get("geometry")
-        if lat and lon and not geometry:
-            try:
-                geometry = {
-                    "type": "Point",
-                    "coordinates": [float(item["lon"]), float(item["lat"])],
-                }
-            except ValueError:
-                logging.warning("Couldn't convert lat (%s) and lon (%s) to float", lat, lon)
-        feature.append(("geometry", geometry))
-
-        return feature
 
 
 class GeoJsonExporter(JsonItemExporter):
