+++ locations/settings.py
@@ -31,8 +31,8 @@ ROBOTSTXT_OBEY = True
 FEED_URI = os.environ.get("FEED_URI")
 FEED_FORMAT = os.environ.get("FEED_FORMAT")
 FEED_EXPORTERS = {
-    "geojson": "locations.exporters.GeoJsonExporter",
-    "ndgeojson": "locations.exporters.LineDelimitedGeoJsonExporter",
+    "geojson": "locations.exporters.geojson.GeoJsonExporter",
+    "ndgeojson": "locations.exporters.ld_geojson.LineDelimitedGeoJsonExporter",
 }
 
 # Configure maximum concurrent requests performed by Scrapy (default: 16)
